#641 Separate font stores to own files
to avoid mega class in PdfBoxFontResolver.
